separate Traversal and PathAwareTraversal
