+++ traversal/src/main/scala/overflowdb/traversal/NodeTraversal.scala
@@ -33,8 +33,8 @@ class NodeTraversal[E <: Node](val traversal: Traversal[E]) extends AnyVal {
   def out: Traversal[Node] =
     traversal.flatMap(_.out)
 
-  def out3: Traversal[Node] =
-    traversal.flatMap3(_.out)
+//  def out3: Traversal[Node] =
+//    traversal.flatMap3(_.out)
 
   /** follow outgoing edges of given labels to adjacent nodes */
   def out(labels: String*): Traversal[Node] =
